Cart can now be emptied with the correct policy in place

+++ app/controllers/carts_controller.rb
@@ -13,6 +13,7 @@ class CartsController < ApplicationController
     if current_order.id != nil?
       session[:order_id] = nil
     end
+    authorize @cart_items
     Order.last.destroy
     redirect_to products_path
   end
